# contracts/http/openapi.yaml
...
openapi: 3.1.0
info:
  title: Research Radar — HTTP wire contract
  version: 0.3.0
  summary: 'Hợp đồng HTTP cho toàn bộ operation `transport: http` của contracts/ports.yaml.'
  description: |-
    Đây là hợp đồng dây (wire) của Research Radar, sinh từ `contracts/ports.yaml` và giữ nguyên `operation_id`
    làm `operationId`. Mỗi operation `transport: http` trong ports.yaml có ĐÚNG MỘT cặp path+method ở đây.

    Những gì file này KHÔNG định nghĩa, và nơi chúng thuộc về:
      - Ngữ nghĩa transition trạng thái  → contracts/state/*.yaml (PC03)
      - Danh mục mã lỗi và envelope      → contracts/errors.yaml (PC03); `ErrorEnvelope` dưới đây là bản sao ĐÚNG
                                           hình dạng của `error_envelope` trong file đó
      - Ngân sách retry/timeout/lease    → contracts/retry-policy.yaml (PC03)
      - Ranh giới module và cạnh bị cấm  → contracts/modules.yaml (PC01)
      - Nội dung/coverage báo cáo        → contracts/reporting/* (PC04)
      - Task và schema AI                → contracts/ai/* (PC06)
      - Ngữ nghĩa lệnh Telegram          → contracts/telegram/commands.yaml (PC07)
      - Vòng đời secret, CSRF, xoay token→ contracts/ops/secrets.md (PC08)

    Wire rules bắt buộc (SRC-PLAN §5.1), áp cho mọi operation:
      - ID truyền dưới dạng string; timestamp UTC RFC 3339 với độ chính xác mili giây (AMD-B08).
      - Mutation từ worker mang `request_id`, `schema_version`, `idempotency_key`, `payload_hash`; mutation đang
...

...
  x-contract:
...
    - REQ-D04
...
servers:
...
security:
- ownerSessionCookie: []
  ownerCsrfToken: []
components:
  securitySchemes:
    ownerSessionCookie:
      type: apiKey
      in: cookie
      name: rr_session
      description: 'Cookie phiên HttpOnly + Secure + SameSite=Lax. PROVISIONAL, phối hợp với PC08 (xem info.x-contract.security_scheme_decision).
        Với MUTATION, scheme này xuất hiện CÙNG `ownerCsrfToken` trong MỘT security requirement — tức quan hệ AND, cookie
        một mình KHÔNG đủ. Với read-only, cookie đứng một mình: CSRF bảo vệ thao tác đổi trạng thái, không phải phép đọc.'
    ownerCsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: 'CSRF double-submit: giá trị phải khớp cookie `rr_csrf` (không-HttpOnly). Bắt buộc cho MỌI mutation của
        owner session. Thiếu hoặc lệch ⇒ 403 `CSRF_REJECTED` (KHÔNG phải `FORBIDDEN_EDGE`: cạnh vẫn hợp lệ, chỉ lời gọi
        không chứng minh được nó do trang của ứng dụng phát ra — ruling CR-PC08-03/CR-PC08-04). Phiên KHÔNG bị hủy khi gặp
        mã này.'
    collectorToken:
      type: http
      scheme: bearer
      description: Token của collector, lưu trong file cấu hình QUYỀN HẠN CHẾ trên máy cá nhân, KHÔNG BAO GIỜ trong repo
        (SRC-SPEC §11.2). Chỉ mở được các operation `worker.*` và `ingest.*`; gọi `save.create` bằng token này ⇒ 401 (denied
        case NC-01).
    analysisWorkerToken:
      type: http
      scheme: bearer
      description: Token của analysis worker. Chỉ mở `analysis.*` và `secret.issue_task_credential`. KHÔNG mở đường tới
        tag, report hay Saved (denied case NC-02).
    telegramIngressSecret:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: Bí mật webhook do app đặt khi đăng ký webhook với Telegram. Sai hoặc thiếu ⇒ 401 và KHÔNG xử lý update.
        Biết chat ID KHÔNG phải là quyền (SRC-SPEC §11.3).
    backupOperatorToken:
      type: http
      scheme: bearer
      description: 'Danh tính operator backup. Tách khỏi phiên owner vì restore là thao tác có khóa side effect (AMD-B11).
        Sửa ở FIX1 (CR-PC08-02): mọi route `backup.*` nhận DUY NHẤT scheme này — trước đây phiên trình duyệt cũng qua được,
        nghĩa là một tab đang mở có thể kích hoạt restore.'
  parameters:
    SchemaVersionHeader:
      name: X-Schema-Version
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/SemVer'
      description: Phiên bản wire schema của request. KHÔNG dùng 'latest' (SRC-PLAN §5). Lệch major ⇒ 422 `VALIDATION_ERROR`
        với `details_safe.violation_kind = schema_version_unsupported`.
    RequestIdHeader:
      name: X-Request-Id
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/Ulid'
      description: ID của LẦN GỌI này. Một retry sinh `X-Request-Id` MỚI nhưng GIỮ NGUYÊN `Idempotency-Key` — đó là cách
        server phân biệt replay với một yêu cầu mới.
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/IdempotencyKey'
      description: Khóa chống trùng do client cấp; phạm vi khóa nằm ở `x-idempotency.key_scope` của từng operation. Cùng
        key + cùng payload_hash ⇒ trả kết quả đã commit. Cùng key + payload khác ⇒ 409 `IDEMPOTENCY_CONFLICT`, KHÔNG ghi
        đè (SRC-PLAN §5.1).
    CsrfTokenHeader:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
        maxLength: 128
      description: Double-submit CSRF; phải khớp cookie `rr_csrf`. Chỉ xuất hiện trên MUTATION dùng phiên owner. Thiếu/lệch
        ⇒ 403 `CSRF_REJECTED`.
...
  schemas:
    ErrorEnvelope:
      type: object
      additionalProperties: false
      description: |-
        Envelope lỗi chung (SRC-PLAN §5.1). Hình dạng này là BẢN SAO của `error_envelope` trong contracts/errors.yaml;
        nếu hai bên lệch nhau thì contracts/errors.yaml THẮNG và file này phải sửa.

        `correlation_id` LUÔN có mặt, kể cả khi `storage.health != healthy`: nó sinh trong bộ nhớ tiến trình, không cần
        một hàng DB. Không mã lỗi nào đòi hỏi ghi được vào DB mới coi là đã báo (contracts/errors.yaml §EPR-01).

        Envelope KHÔNG BAO GIỜ chứa: transcript của model, prompt gốc, nội dung post nguyên văn, API key, bot token,
        cookie, header Authorization, chat ID, đường dẫn file tuyệt đối, stack trace, câu SQL, tên bảng/cột nội bộ.
      required:
      - code
      - scope
      - retry_class
      - message_safe
      - correlation_id
      properties:
        code:
          type: string
          enum:
          - X_CHALLENGE_REQUIRED
          - X_ACCESS_BLOCKED
          - WORKER_LEASE_EXPIRED
          - STALE_LEASE
          - INGEST_ACK_LOST
          - IDEMPOTENCY_CONFLICT
          - SOURCE_METADATA_UNAVAILABLE
          - IDENTITY_CONFLICT
          - AI_OUTPUT_INVALID
          - AI_PROVIDER_UNAVAILABLE
          - AI_ATTEMPT_UNCERTAIN
          - EMBEDDING_GENERATION_MISMATCH
          - TAG_VERSION_STALE
          - CONFLICT
          - TELEGRAM_SEND_UNCERTAIN
          - TELEGRAM_PERMANENT_FAILURE
          - UNAUTHORIZED_COMMAND
          - STORAGE_WRITE_FAILED
          - RESTORE_UNVERIFIED
          - SOURCE_LAYOUT_CHANGED
          - CSRF_REJECTED
          - UNAUTHORIZED
          - CAPABILITY_DENIED
          - FORBIDDEN_EDGE
          - VALIDATION_ERROR
          - NOT_FOUND
          - RATE_LIMITED
          - INTERNAL
          description: Phải là một mã đã khai trong contracts/errors.yaml. Không có mã ngoài danh mục.
        scope:
          type: string
          enum:
          - run
          - item
          - delivery
          - storage
          - request
          description: Bằng đúng `scope` của mã trong contracts/errors.yaml.
        retry_class:
          type: string
          enum:
          - none
          - retryable_with_budget
          - needs_user
          - operator_decision
          - unknown_outcome
          description: Bằng đúng `retry_class` của mã. `unknown_outcome` nghĩa là KHÔNG BIẾT side effect đã xảy ra hay chưa
            — client KHÔNG được coi là rollback.
        message_safe:
          type: string
          maxLength: 300
          description: Tiếng Việt, render từ `message_safe_template` của mã. KHÔNG chèn giá trị ngoài `details_safe_keys`.
        correlation_id:
          $ref: '#/components/schemas/Ulid'
        details_safe:
          type:
          - object
          - 'null'
          description: Chỉ chứa các khóa liệt kê ở `details_safe_keys` của mã trong contracts/errors.yaml. Khóa lạ ⇒ producer
            tự từ chối, không gửi ra.
        retry_after_ms:
          type:
          - integer
          - 'null'
          minimum: 0
          description: Chỉ có mặt khi `retry_class = retryable_with_budget` hoặc mã `RATE_LIMITED`. null nghĩa là KHÔNG
            có gợi ý, KHÔNG phải 'retry ngay'.
    Ulid:
      type: string
      pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
      description: ID truyền dưới dạng string (SRC-PLAN §5.1).
    SemVer:
      type: string
      pattern: ^[0-9]+\.[0-9]+\.[0-9]+$
...
    IdempotencyKey:
      type: string
      pattern: ^[A-Za-z0-9_.:-]{16,128}$
...
    GenericObject:
      type: object
      description: |-
        Chỗ giữ có chủ đích: hình dạng chi tiết thuộc một gói khác và CHƯA tồn tại tại thời điểm viết file này.
        Xem `$ref` trong `x-response-schema-planned` của ports.yaml cho từng operation. Đây KHÔNG phải giấy phép chấp
        nhận trường tùy ý: gói sở hữu phải thay bằng schema đóng với `additionalProperties: false` trước G3.
...
paths:
...
  /v1/settings:
    get:
      operationId: settings.get_config
      summary: Đọc cấu hình (lịch, giới hạn, provider/model theo tác vụ, embedding, liên kết Telegram)
      description: CHỈ trả THAM CHIẾU tới secret, không bao giờ giá trị key (denied case NC-03).
      tags:
      - settings
      security:
      - ownerSessionCookie: []
      x-auth-scope: owner_session
      x-owner-module: MOD-settings-service
      x-caller-modules:
      - MOD-web-ui
      x-mutation: false
      x-scenario-refs:
      - SC17
      x-requirement-refs:
      - REQ-D39
      - REQ-D56
      - REQ-P0-11
      x-decision-refs:
      - B08
      - B13
      x-invariant-refs:
      - I01
      - I11
      parameters:
      - $ref: '#/components/parameters/SchemaVersionHeader'
      - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          description: Cấu hình đầy đủ với secret ở dạng **tham chiếu + trạng thái đã cấu hình**, không bao giờ là giá trị
            key (SRC-SPEC §11.2). Kèm timezone IANA đang dùng (B08).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
          headers:
            X-Schema-Version:
              description: Phiên bản schema của response.
              schema:
                $ref: '#/components/schemas/SemVer'
        '401':
          description: 'Lỗi: `UNAUTHORIZED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 'Lỗi: `FORBIDDEN_EDGE`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: 'Lỗi: `INTERNAL`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      operationId: settings.update_config
      summary: Sửa cấu hình
      description: 'Delta cấu hình: lịch, timezone IANA, giới hạn mỗi đợt, N ngày với ngược, provider/model theo tác vụ,
        key mới (chỉ ghi vào secret service), model embedding.


        Ghi bản cấu hình mới; đổi model embedding kích hoạt embedding.start_generation_rebuild (không tự chuyển active generation);
        key mới đi qua secret.store_provider_key.'
      tags:
      - settings
      security:
      - ownerSessionCookie: []
        ownerCsrfToken: []
      x-auth-scope: owner_session
      x-owner-module: MOD-settings-service
      x-caller-modules:
      - MOD-web-ui
      x-mutation: true
      x-scenario-refs:
      - SC24
      x-requirement-refs:
      - REQ-D39
      - REQ-D40
      - REQ-D48
      - REQ-P0-11
      x-decision-refs:
      - B08
      - B13
      - B14
      x-invariant-refs:
      - I01
      - I12
      - I14
      x-idempotency:
        required: true
        key_scope: request_id
        rule_vi: Cùng request_id + cùng payload_hash trả bản ghi đã commit; payload khác trả IDEMPOTENCY_CONFLICT.
        header: Idempotency-Key
      parameters:
      - $ref: '#/components/parameters/SchemaVersionHeader'
      - $ref: '#/components/parameters/RequestIdHeader'
      - $ref: '#/components/parameters/IdempotencyKeyHeader'
      - $ref: '#/components/parameters/CsrfTokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
      responses:
        '200':
          description: Cấu hình sau khi ghi, secret vẫn ở dạng tham chiếu.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
          headers:
            X-Schema-Version:
              description: Phiên bản schema của response.
              schema:
                $ref: '#/components/schemas/SemVer'
        '401':
          description: 'Lỗi: `UNAUTHORIZED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 'Lỗi: `CSRF_REJECTED`, `FORBIDDEN_EDGE`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 'Lỗi: `IDEMPOTENCY_CONFLICT`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 'Lỗi: `VALIDATION_ERROR`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 'Lỗi: `RATE_LIMITED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            Retry-After:
              description: Giây do nguồn/API chỉ định. Giá trị này THẮNG mọi backoff cấu hình.
              schema:
                type: integer
                minimum: 0
        '500':
          description: 'Lỗi: `INTERNAL`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 'Lỗi: `STORAGE_WRITE_FAILED`. `storage.health = write_blocked` hoặc `maintenance`: mutation KHÔNG
            được ACK và KHÔNG có gì được ghi (SRC-PLAN §8.4, I02).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            Retry-After:
              description: Giây. Gợi ý theo `storage_recovery_probe` (30 s).
              schema:
                type: integer
                minimum: 0
  /v1/settings/providers/{provider_name}/test:
    post:
      operationId: settings.test_provider
      summary: Thử một provider AI đã cấu hình (đường API key, phía server)
      description: 'Provider ID cần thử; không nhận key trực tiếp từ frontend.


        Ghi provider_test_result mới nhất; không đổi cấu hình.


        Kết quả `usable | unusable | unknown` + reason_code; `unknown` KHÔNG được ghi thành 0 hay thành `unusable` (I14).'
      tags:
      - settings
      security:
      - ownerSessionCookie: []
        ownerCsrfToken: []
      x-auth-scope: owner_session
      x-owner-module: MOD-settings-service
      x-caller-modules:
      - MOD-web-ui
      x-mutation: true
      x-scenario-refs:
      - SC16
      x-requirement-refs:
      - REQ-D41
      - REQ-D51
      x-decision-refs:
      - B13
      x-invariant-refs:
      - I11
      - I14
      x-idempotency:
        required: false
        key_scope: null
        rule_vi: Mỗi lần thử ghi một bản ghi kết quả mới; không dùng lại kết quả cũ.
        header: null
      parameters:
      - name: provider_name
        in: path
        required: true
        schema:
          type: string
          maxLength: 64
        description: Tên provider đã cấu hình.
      - $ref: '#/components/parameters/SchemaVersionHeader'
      - $ref: '#/components/parameters/RequestIdHeader'
      - $ref: '#/components/parameters/CsrfTokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
      responses:
        '200':
          description: Kết quả `usable | unusable | unknown` kèm lý do đã che secrets. Đường CLI/ACP **không** thử được
            từ server; trạng thái của nó chỉ đến từ worker.register_capabilities trên máy cá nhân.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
          headers:
            X-Schema-Version:
              description: Phiên bản schema của response.
              schema:
                $ref: '#/components/schemas/SemVer'
        '401':
          description: 'Lỗi: `UNAUTHORIZED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 'Lỗi: `CSRF_REJECTED`, `FORBIDDEN_EDGE`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 'Lỗi: `NOT_FOUND`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 'Lỗi: `VALIDATION_ERROR`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 'Lỗi: `RATE_LIMITED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            Retry-After:
              description: Giây do nguồn/API chỉ định. Giá trị này THẮNG mọi backoff cấu hình.
              schema:
                type: integer
                minimum: 0
        '500':
          description: 'Lỗi: `INTERNAL`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: 'Lỗi: `AI_PROVIDER_UNAVAILABLE`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...